Add local-buffer-wrapper class
MemFd buffers don't need to be downloaded into host memory, because they
are already available in host memory. This also means, that they don't
need to be duplicated.
Add a class, that wraps RDP buffers, which use a MemFd buffer
underneath. Since this class is derived from the abstract local-buffer
class, the software-based RFX encode-session will be able to use these
local-buffer-wrapper classes without the need of adjusting.