Add generic damage detector class
Before attempting to encode a frame, gnome-remote-desktop performs
damage detection to only encode damaged frames, to decrease the update
region for the client, and if possible to tell the encode-session to
only consider a specific region.
When using hardware acceleration, the damage detection process happens
at the same time as the colour conversion process. The reason for this
is that global memory access in shaders is the most slowest action of
all possible actions. So, in shaders a pixel value is only read once,
then used for comparison and for conversion.
This situation does not apply for software rendering, as the
encode-session performs the colour conversion itself. Also, it would be
hard to apply SIMD actions here. So, for software rendering, but only
software rendering a separate damage detection handling is needed.

Add a class that handles damage detection for frames, that lie in host
memory. Don't use the existing damage-detector classes, as they are
incompatible with the current damage handling:
The damage-detector classes use one uint8 for each 64x64 damage area.
This new class uses one uint32. The reason is compatibility with the
accelerated paths, as all GPUs support 32-bit integers for buffer
memory, but not 8-bit integers for buffer memory.
Also, since this new class only affects software rendering, it does not
need all the abstraction in the damage-detector classes.
This class is generic, as it is unrelated to any codec in its handling.
It will be used by both dma-buf- and MemFd-path, when software encoding
needs to be done.