rdp: Add generic software-based view-creator
This view-creator works similarly to the generic view-creator using GL.
The main difference here is though that no dma-buf image needs to be
downloaded. Since the create_view() function is invoked on the graphics
thread, do not launch the damage detection handling there, as it is not
an async action. Instead, launch it in the finish_view() function, like
in the other view-creators, as this function is invoked in the
respective view-creator thread, which thus does not block the graphics
thread.
With MemFd buffers, there is another difference: The used RDP buffers
will still be needed after the view creation, as the colour conversion
in the software-encoding path is done by the encode-session and the
buffers don't need to be downloaded. This issue will however be
addressed in the next commits.