rdp: Partially enable new render path
With the fallback render handling for the new render path now being
implemented, make use of the new render path, when no NVIDIA GPU is
used. Since the new MemFd handling won't involve a buffer copy anymore,
this also reduces the frame latency for the MemFd path.
As long as the hardware accelerated handling for NVIDIA GPUs in the new
render path is not done, the legacy render handling is still needed.
When that's done, the legacy render path can be removed.

While the new render path now be used for software encoding, don't
enable yet the VA-API encoder by default. For this, one last missing
piece is still needed, and that is explicit sync support.