rdp-frame: Delay view finalization when needed
When using software encoding and MemFds, the software-based
encode-session performs the colour conversion itself and the content of
the MemFd buffers does not need to be duplicated. Currently, the
reference count of the RDP buffers, that are used for the view-creation,
are unreferenced when the view has been created. For the MemFd path,
however, this won't work, as the source RDP buffer for the frame is
still needed for the encode-session.

To fix this problem, use the previously added API to ask the
render-context, whether the view-finalization (which involves
unreferencing the source buffers) needs to be delayed. In that case, the
view is finalized before the frame struct is destroyed.