rdp: Add class for handling multitouch input events
Touch (and (multi-)pen) input is not part the RDP core protocol.
Instead, they are transmitted via the Input Virtual Channel Extension.
So add a class for this DVC and derive it from the GrdRdpDvc.
Compared to "normal input" (keyboard and mouse events), touch input in
RDP is state based. This means, that input events cannot just be
submitted to the RDP event queue. Instead, for each touch contact, the
resulting event is based on the state the touch contact is in, and the
contact flags, which determine the next state of the touch contact. This
state diagram is documented in 3.1.1.1 Touch Contact State Transitions
([MS-RDPEI]).
To not leak any [MS-RDPEI] specific types and internals to the RDP event
queue, handle these touch contact transitions inside this new DVC class.

Based on the type of the touch contact id (uint8), 256 simultaneous
touch contacts are possible. So add an array with 256 touch contexts to
this class.
Each touch context tracks the state of the automata in 3.1.1.1
([MS-RDPEI]) for the respective touch contact, a GrdTouchContact object,
and a flag, whether the contact should be ignored. The GrdTouchContact
object here is needed for the previously added API in GrdSession to
update a touch contact.
The ignore_contact flag is used to track whether a touch contact should
be ignored. This flag is used, when the touch position cannot be
transformed. Usually, that is the case, when the client sends a
position, that is outside of any (virtual) monitor here. The client does
not necessarily know, whether a touch contacts position cannot be
transformed and thus considers the contact to be valid. This is not a
problem, as the event can simply be ignored here. However, the follow-up
events depend on the contact state here. Thus, continue to handle the
contact flags of the event to transition the touch contact state in the
next state. When the original state ('Out of Range') is reached, reset
the ignore_contact flag.
When an input event is received, do not directly handle it. Instead,
push it to an internal queue, which is processed, just like the ei
source in GrdSession, on the main thread. This ensures, that no race
conditions happen, as ei touch contacts can only be acquired, when there
is an ei touch device.
Currently, only touch events are handled and not pen events, as libei is
missing the needed API for it. Also, in RDP, touch events can contain an
optional (rotated) contact rect and a value describing the pressure of
the touch contact. Discard these optional parameters, as there is no
respective libei API for them too.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/314>