session: Add ei touch device
In order to be able to submit touch input events to mutter, a respective
ei device needs to be added. So add it.
Like absolute pointer motion events, touch input events are absolute
motion events, so also import the respective ei regions too. However,
use a separate hash table for this, as the same ei region can be used
for multiple devices. Otherwise, the respective GrdRegions of another
absolute motion device could be accidentally be replaced, as the
mapping-id of the ei region is used as the hash table key.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/314>