session: Emit signals when adding or destroying the ei touch device
When an ei touch device is destroyed, the handler of touch events needs
to be notified to reset all contact states. To be able to do that, add a
signal to signal the touch event handler, that the respective ei touch
device has been destroyed.
In addition to that, add a signal to notify the touch event handler,
when an ei touch device was added, allowing the touch event handler to
start processing touch events.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/314>