rdp-dvc-input: Reset touch contexts when touch device is recreated
When the monitor layout of an active remote desktop session changes,
mutter recreates all ei regions and thus all ei touch devices too.
This makes sense, as the input regions changed. In this situation, all
active touch contacts become invalid and thus need to be destroyed.
On the side of the remote desktop client, the same seems to happen: The
internal state machine for each contact is reset to the 'Out of Range'
state.
The handling for this on the server side can actually be combined with
the recreation of the ei touch device. To do that, hook up the
previously added signals in GrdSession, that notify connected handlers,
that the ei touch device was destroyed. To handle a previously active
contact, simply release the GrdTouchContact and reset the state of the
[MS-RDPEI] specific contact automata.

During testing, a bug with Microsoft Windows App on the iPhone was
found: When a touch contact was held (touch-down) and the screen
rotated, then the client resets the automata states. However, for some
reason, it continues to send spurious touch-motion events with the same
position over and over again, even when the touch contact is not held
anymore or the finger moved to a different position while the touch
contact is held.
When performing a touch-down again, these spurious touch-motion events
are not sent anymore. The client here also sends the contact flags for a
touch-down event, meaning that the touch-motion events were not
intended.
To avoid unnecessary warnings, add a workaround for this client bug on
the server side: When the ei touch device is recreated and all related
touch contacts are reset, set the ignore_contact flag, when the contact
is not already in the 'Out of Range' state. In addition to that, reset
the ignore_contact flag, when handling the 'Out of Range' state with a
valid contact flag combination for that state.
This gets rid of a spam wave of these spurious touch-motion events,
while keeping protocol conform behaviour.

When an ei touch device is added, trigger the event source to ensure,
that events, that were received before the new ei touch device was
added, are actually handled. Use for this the previously added
'touch-device-added' signal.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/314>