session: Use separate libei-regions hash table for abs pointer
Currently, the regions hash table uses for its key the mapping-id of a
stream and the respective region as its value. This allows a quick
lookup of the region for a specific stream.
However, streams use the same mapping-id for each device. For input
transformation this makes sense, but the associated device for a region
is not the same, as the required one.
This becomes a problem, as soon as touch support is added, as the
regions for the touch device would override the regions for the absolute
pointer device, as both use the same mappings-ids. This effectively
makes absolute pointer motion events impossible.

Fix this problem by using a separate hash table for each
absolute-motion-based device. In future commits, the touch device will
use its own hash table too.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/314>