build: Bump required dependency versions for multitouch support
In order to be able to use the input DVC for multitouch input, FreeRDP
3.15.0 or later is required, as older versions have a bug, that the DVC
does not open correctly. Also, the event handling was not done
automatically in older versions.
In addition to that, ei_touch_cancel() is needed in order to be able to
cancel touch events. This, however, requires libei 1.3.901 or later.
So bump the required dependency versions accordingly.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/314>