daemon-system: Consider remote sessions from autologin
GDM daemon can start a remote user session from the grd system daemon.
Now, it can start a remote user session independently too, from a GDM tool
which uses the autologin service.

In the autologin process, two remote displays will be created
sequentially:
  1. A first remote display with no session_id and remote_id: "/".
  2. A second remote display with a user session_id and a remote_id
     inherited from the first one.

For the first display, create an empty remote_client, generate a
remote_id and set it to the display calling "SetRemoteId".

Then the second display will be created, which has a session_id and a
remote_id, so a handover interface will be registered to allow connecting
with a potential handover daemon from a remote login.

There could be the case where both displays have as remote_id: "/".
This means that two remote clients will be created and "SetRemoteId"
will be used on both of them:
  - The first remote display won't have any handover interfaces, so when
    removed, its remote_client will be removed.
  - The second remote display will have one handover interface, so it will
    be available for a potential handover.