utils: Fix toggle_systemd_unit
This function is called from a pkexec spawned child of grdctl. This
child has the configure-system-daemon polkit permission.

The toggle_systemd_unit spawns the tool grd-enable-service. This tool
will only work if its caller has the configure-system-daemon polkit
permission. It checks that from the pid argument it receives.

The pid it was getting was the pid of parent that called pkexec which doesn't
have the authorization. Fix it providing the right pid.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/320>