data: Update configure-system-daemon.policy
When grdctl is configuring the system daemon, it needs to get
configure-system-daemon polkit permission. This should get it when it's
spanwed with pkexec.

However this rule is also needed to get the permission, otherwise it
won't have the permission and won't be able to enable/disable
the systemd service.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/320>