Introduce throttler class
This handles the throttling of incoming connections, and is intended to
act as a first line of defence against denial of service attacks. It
implements the following:

 * It limits the global number of concurrent active connections
 * It limits the number of concurrent active connections per peer
   (roughly IP address)
 * It limits the number of new connections per second coming from a peer
 * It maintains a limited number of pending connections that is waiting
   to be handled

This avoids running into situations where an uncontrolled number of
connection attempts causes resources to be exhausted, e.g. number of
file descriptor exceeding the process limit.

The limits are currently hard coded.

Related: CVE-2025-5024
Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/321>