daemon-handover: Only logout when system daemon vanishes for greeters
Previously, if the system daemon was restarted, it didn't have any way
of reconnecting to previous handover daemons, so the remote sessions
were left orphaned without a way of connecting to them. Now it's changed and
will register handover interfaces again.

This can only work for user remote sessions, to perform a handover from
a login session. It can't work for login sessions. In that case, these
login sessions will be terminated.