hwaccel-vulkan: Add API to get DRM format modifiers for DRM format
Graphics drivers might not advertise the same DRM format modifiers for
Vulkan and EGL. In order to be able to fetch the supported DRM format
modifiers in Vulkan, add an API for it.
Despite supporting Vulkan, it might later still not be clear whether
Vulkan can be used for a PipeWire stream, or whether EGL will still need
to be used that.
To account for that scenario, always return a common set of supported
DRM format modifiers that are both supported by Vulkan and EGL.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/324>