rdp: Get DRM format modifiers from Vulkan if needed
When a GPU supports both Vulkan and VA-API, gnome-remote-desktop tries
to encode a PipeWire stream using AVC, if the RDP client supports that.
But in order to import the dma-buf images, gnome-remote-desktop always
uses the DRM format modifiers advertised via EGL, although Vulkan is
used in that case. In theory, the advertised DRM format modifiers via
EGL might not be the same as the ones advertised via Vulkan.

Fix that issue by using the previously implemented API to fetch the DRM
format modifiers from Vulkan when Vulkan is available. As a common set
for Vulkan and EGL is returned, these DRM format modifiers should always
work. When Vulkan is not available, use the DRM format modifiers from
EGL as done before.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/324>