daemon-handover: Cancel dbus operations on shutdown
When shutting down the handover daemon, the remote desktop handover
object is cleared. However, when there are still dbus operations
pending, then these objects still exist due to associated GTasks.
While these GTasks got a GCancellable object, this cancellable is not
cancelled when the handover object is cleared.
So, cancel it manually on shutdown. With the shutdown of the handover
daemon any possible pending handover action is dropped, thus some error
messages appear on the journal, which have no relevance. Do not emit
these error messages in this case by checking whether the respective
cancellable of a GTask was cancelled.

https://errors.ubuntu.com/problem/3e3d1f4e8b3350537a134db3edd574f99332a6b5

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/325>