daemon-handover: Stop dereferencing potentially invalid parameters
The handover daemon calls several dbus async calls, which may outlive
the handover daemon itself. In that case, the user data pointer becomes
invalid. However, in a lot of cases, it is actually the one that is
dereferenced first and thus leading to a potential crash.

Fix this issue by always calling the respective finish() function first.
Only when that succeeds, the user data can be deemed as valid, as the
finish() function will fail when the respective Cancellable was
cancelled before.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/325>