rdp-renderer: Move GError into the right scope
The GError is not only just used inside the for-loop. It could
potentially be set again when it was already set. In such case, glib
prints a warning because the logic is just wrong.
So, fix this issue by simply moving the respective GError inside the
for-loop. Due to it being a g_autoptr, it should be cleared upon leaving
each for-loop run.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/331>