rdp/rdpgfx: Fix black screen after second graphics pipeline reset
When mstsc or msrdc encounter an error in the graphics pipeline that
they deem as server error, they don't directly error out. Instead, they
try to emit a graphics pipeline reset by sending the capability sets
again. However, after gnome-remote-desktop reset the graphics pipeline,
only a black screen is visible. The reason for this is that no rendering
happens. And the reason for that is, that the frame controller still
expects some frame acks to be delivered, although the whole protocol is
reset.
In addition to this problem, gnome-remote-desktop currently silently
removes all surfaces from the surface table despite them still being
there. These surfaces are still needed in order to be able to reset the
behaviour induced by the attached frame controllers.

Fix this issue by replacing the silent surface steal operation with an
operation that resets all frame controllers.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/331>