rdp-monitor-config: Stop requiring even widths for monitor sizes
2.2.2.2.1 DISPLAYCONTROL_MONITOR_LAYOUT ([MS-RDPEDISP]) explicitly
states that the widths must not be an odd value. However, mstsc and
msrdc don't refrain from sending such values when setting such value in
the respective RDP settings file. MS Windows RDS handles such situations
by rounding up the value. gnome-remote-desktop currently rejects such
monitor configurations. The original intention was to be able to
decrease the GPU time in the AVC CUDA kernel by reading two pixels as
one uint64 instead of two uint32s, thus increasing performance due to
lesser global memory access.
However, the CUDA path in gnome-remote-desktop is expected to be
replaced with Vulkan Video, in which case the view shader is shared with
the one for VA-API. As that shader reads from an image instead of an
image, disguised as a buffer, this optimization won't be needed anymore.
So, drop the even-width enforcement and allow odd monitor sizes.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/333>