encode-session-vaapi: Align source surface width to the value 32
2.2.4.4 RFX_AVC420_BITMAP_STREAM ([MS-RDPEGFX]) demands that the source
width and height of the AVC bitstream are aligned to the value 16. This
is what gnome-remote-desktop does and so far that works fine, except for
msrdc: When running msrdc, then the client errors out when using AVC for
some monitor sizes, e.g. 1697x930. The aligned size is here 1712x944,
which is though aligned to 16x16 tiles. While investigating this issue,
it seems that msrdc does not error out when aligning the width to the
value of 32. The height can remain aligned to the value 16. With this
behaviour, msrdc does not error out when a bitstream for a monitor with
a size of e.g. 1697x930 is received.
While investigating this issue further with Martin Fleisz (FreeRDP,
Thincast) by comparing the behaviour with the RDP server from the Ogon
project, it became clear that msrdc's behaviour depends on whether
hardware acceleration is available on the client side or not. With
hardware acceleration, an alignment to 16x16 tiles was enough, just like
the documentation demands. However, when running MS Windows inside a
virtual machine, hardware acceleration was not available anymore and the
software decoder was used, which apparently requires an alignment to
32x16 tiles apparently.

Despite the error being in Microsofts implementation, align the source
surface width to multiples of 32 to cope with this problem. Microsoft is
known to not fix bugs in their implementations and due to a heavy use of
standard resolutions in Microsoft RDP implementations, their developers
likely did not encounter that issue so far.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/333>