tests/rdp: Ensure server instance runs headlessly
The RDP test not only tests whether connecting to gnome-remote-desktop
works and getting a fullscreen update, but also the virtual monitor
functionality, that the requested monitor size complies with the actual
monitor size.
In the next commit, gnome-remote-desktop receives support for extending
the current desktop with a virtual monitor. That mode will use embedded
cursors. These cursors can lead to updates, where the initial
invalidation rectangle on the FreeRDP client side does not correspond to
the actual monitor size. The invalidation rectangle is, however, used
for the monitor size comparison.
The different invalidation rectangle does not mean that the requested
monitor size is wrong. FreeRDPs client handling somehow seems to emit
different invalidation rects in this situation.

To be able to distinguish between these situations and to ensure that
the original intention of the invalidation rect on the FreeRDP client
side is met, always use the headless mode, as in this mode, the
invalidation rect always seems to meet the requested virtual monitor
size.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/334>