rdp: Ignore PipeWire stream resize events with invalid sizes
When quickly resizing a virtual monitor, every resize event will go
through the state machine of the layout manager. However, due to a
recent bug in either mutter or PipeWire, the PipeWire stream size will
not be directly updated. Instead of one resize update with the new size,
gnome-remote-desktop seems to receive one resize event with the old size
again and only then another resize event with the correct size.

So, instead of treating the first wrong resize size as a fatal error,
simply ignore it and wait for the next resize event. This way the remote
session won't directly be stopped due to wrong behaviour, as with some
clients, this bug can be hit quite easily.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/336>