rdp-render-context: Disable VA-API when using an AMD GPU
For some reason, AMDs VA-API driver always creates bitstreams, that are
just a few bytes large, containing not even enough bytes for the
H.264-SPS and PPS headers, while the encode operations never "fail", as
the driver always reports that no error happened.
As the client side will not be able to decode the bitstream and thus see
a frame, simply disable encoding with VA-API when an AMD GPU is used.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/362>