rdp-view-creator-avc: Fix too large shader invoke group size
The compute shader for the view creation has local-size/warp-size of
16x16. This means that the amount of warps/groups to invoke needs to be
divided by 16 for the width and 16 for height. However, each streaming
multiprocessor itself works on a 2x2 field in order to handle the
subsampling operations. So, both the width and height need to be divided
by 2 again. This is currently not the case, meaning gnome-remote-desktop
will invoke the view creation shader for 4 times its size, which is
wasteful regarding compute resources.

Fix the group invocation size by also dividing the size by two. On a
Haswell iGPU, this decreases the needed computation time by 50%.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/336>