rdp-pipewire-stream: Use dynamic spa builder
This eliminates any risk of stack overflow when building spa objects,
since the builder dynamically allocates the parameters. It does mean we
need to keep track of the offsets, not the pointers, of the params, as
the dynamically growing builder may move the allocated memory.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/379>