Add frame clock class
When creating the producer side of a camera stream, clients (the camera
stream consumers) expect that the producer side drives the stream. For
this, some kind of frame clock is needed that triggers every time a
frame is needed for a given framerate.
For the trigger mechanism use timer fds, which trigger in the wanted
interval, and combine that with a GSource, that also triggers, when the
timer fd can be read. In such case, call the provided callback function.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/360>