Add helper for encapsulating sample buffers
This sample buffer will be used for camera streams later in order to
save the received sample data from clients. The data buffer itself may
be allocated, but it can also be mapped memory for usage in the host
space. This structure abstracts that aspect away.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/360>