Add abstract decode session class
Camera streams usually contain encoded bitstreams, that need to be
decoded. The scope for gnome-remote-desktop is to have hardware
accelerated decoding later for camera streams. However, not every
hardware supports that and thus a software fallback is needed. In order
to have a unified interface to work with, add an abstract decode session
class, which will be used by every actual decode session implementation.

The decoding operations are intended to run independently from the
actual camera handling itself, also in order to not block the actual
stream handling. Thus, run the actual decoding operation in a separate
thread.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/360>