rdp-surface-renderer: Generalize queued buffer unref operation
The origin of the buffer unref operation does not necessarily happen on
the graphics thread, which is why its unref operation is queued onto the
graphics thread. The used RDP buffers are, however, not the only
objects that need to be unreffed on the graphics thread. So, generalize
this unref queue behaviour in order to be able to handle other objects
in the future.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/356>