rdp: Queue render-context release instead of immediately releasing it
The lifetime of all frame related objects needed to render a frame is
bound to such a frame. This also includes a reference to the
render-context, as all frame related operations depend on it. Thus, when
shutting down the renderer, the renderer waits until all acquired
render-contexts are released. As a result of that, the render-context
release is the last operation when freeing up the resources of a frame.
However, the used RDP buffers for a frame are not directly cleared, but
instead their unref operation is queued onto the graphics thread,
resulting in clearing the render-context before unreffing the respective
buffer. When using software rendering, this can result in an assertion
hit when stopping the RDP session due to the still referenced RDP
buffer, as when using software rendering, the used RDP buffers are not
unreffed after the view creation, but after the encoding operation.
To be precise, gnome-remote-desktop hits the assertion
g_assert (g_hash_table_size (surface_renderer->acquired_buffers) == 0);
in grd_rdp_surface_renderer_invalidate_surface_unlocked(), when tearing
down the graphics pipeline. An action that happens after the render is
shut down.

To fix this issue, do not immediately release the render-context when
clearing an RDP frame. Instead, queue its operation. Since the buffer
unref operations are queued before the render-context, the overall
order that the render-context is released as last is preserved.

See also: https://errors.ubuntu.com/problem/671abb4ed4d2d1e158e89428519dfa20f9218778

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/356>