session: Fix race condition between ready and started signal
The event handling of the ready and started signal, especially in the
RDP backend, assumes that the ready signal is always called before the
started signal. With commit 6153f1920db5ada529455d4c38681f595876347f,
this assumption is now not necessarily true anymore. And so it is now
possible that the started signal is called before the ready signal.
In order to fix this issue, delay the started the signal until after the
ready signal in case the started signal emitter is handled before the
ready signal emitter.

Fixes: 6153f1920db5ada529455d4c38681f595876347f
Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/356>