rdp/rdpecam: Also fetch device properties on device initialization
Version 2 of the camera redirection protocol described in [MS-RDPECAM]
allows servers to also control the camera. This includes settings like
the camera zoom, brightness, etc..
Currently, PipeWire does not provide the ability for PipeWire streams to
provide control for these properties. However, this may change in the
future and for debug purposes having the ability to inspect the
supported device properties can already be helpful to support the effort
for a future camera property handling for PipeWire streams. So, add that
inspect handling.
The inspect handling itself works similarly to the media type
inspection.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-remote-desktop/-/merge_requests/360>